requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
deep-translator==1.11.4
requests==2.32.3
python-dotenv==1.0.1
